nvmf: Fixes outstanding io counting for NVME verify
The is_aer check doesn't check for the adminq to determine
if the cmd is an async event request. This can clash with
the verify IO cmd that has the same opc 0xc as the
async event request.
In the case of an nvme verify, io_outstanding would not
be decremented which then prevents the subsystem from ever
being paused.

Added adminq check for is_aer determination.

Change-Id: I498e85f761258c331ae4c206813a802bbea786d1
Signed-off-by: Michael Haeuptle <michael.haeuptle@hpe.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26068
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>