memory: add helper for walking over non-2MB aligned regions
Iterating over a non-2MB aligned memory regions is going to be useful in
any place that supports 4KB mappings.  This function chooses page size
based on the alignment of the address and doesn't check the map to see
if a 2MB-aligned region actually uses 4KB translations.

Change-Id: Iddb93dc5142de10a553c660461f6085aaa6679f1
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26215
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot