bdevperf: store md_buf size in struct bdevperf_job
We're going to need it in a few more places in the next patch.

Note that we use spdk_bdev_get_md_size() in place of
spdk_bdev_desc_get_md_size() to allocate task->md_buf.  This is fine,
because we first check spdk_bdev_desc_is_md_separate() in which case
both functions return the same value.

Change-Id: Ib36303cc9a83025efd61bbbe1d93c9a5496aa019
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26225
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot