lib/thread: add spdk_io_channel_ref for taking ref of existing channels
There are cases that we just want to take a reference for an existing io
channel. An existing example is that bdev reset requests take extra
reference to bdev io channels to protect the io channels during reset
operations.

A more universal case is to take reference to an io channel for each
request sent to it to protect the channel during io operations, so that
users do not need to implement their own refcounting mechanisms.

Therefore, a light-weight and lockless alternative to spdk_get_io_channel
could be useful.

Change-Id: I570fe47f0ad7fb4943fc20af6d76aadb37e51d86
Signed-off-by: Jinlong Chen <chenjinlong.cjl@alibaba-inc.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26232
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Ben Walker <ben@nvidia.com>