python/cli/nvmf: Make admin cmds passthru a list parameter
Currently SPDK supports setting identify_ctrlr admin cmd
as passthru as separate argument in CLI.
Support for more admin cmds is currently under development.
Having them as separate arguments would add a lot of parameters.
To avoid that - change this single on/off flag parameter to
a list of admin cmds that user would like to be set as passthru.

Change-Id: Ibda375319674794878607bb92d5f45f327f537fe
Signed-off-by: Pawel Baldysiak <pawel.baldysiak@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26236
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>