fix:SPDK acts as a SCSI target, a host performs the following operations on a specific IT nexus:
1. Register & Ignore Existing Key operation, SAK=0x0a, result OK

2. Reserve operation, RK=0x0a, result OK

3. Register & Ignore Existing Key operation, SAK=0x0b, RK=0x0a, result
OK

4. Read Reservation operation - the expected query result should be
RK=0x0b, but the actual result is RK=0x0a, which doesn't match
expectations.

When a Register operation or Register & Ignore Existing Key operation
changes the RKey of an already registered nexus, if that registered
nexus is simultaneously the holder of the reservation, SPDK fails to
change the holder's CKey. This leads to the Read Reservation operation
retrieving the old RKey.