lib/nvme: use changed attached namespace list
Previously the buffer from the log page went unused, while we could limit the
number of identified namespaces to just that list.

The log page contains namespaces which were just made active or were resized,
those are exactly the ones we want to identify.
Inactive namespaces do not need to identified and are already removed from
the list of active namespaces after identify to the controller was sent,
skip those.

Change-Id: I406ed3969d560b79c0ed848815d99b5d692ca37d
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26305
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>