nvmf: Add support for get_supported_log_pages log page
This commit adds support for optional log page
SUPPORTED_LOG_PAGES that gathers all
log pages that are supported by subsystem.
Support is added for both discovery and IO subsystems.

Add public call to get just the stucture filled with what
is supported in SPDK for the controller - it can be used
in future in fixup for passthru get_log_page handling.

One of test_get_log_page()'s case expect LID 0x0 to fail
as not supported. Since it is supported now - change
LID in request to -1.

Change-Id: I30615600930d2210601e92b57eaf02e3e8f8bf07
Signed-off-by: Pawel Baldysiak <pawel.baldysiak@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26230
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>