nvmf: Select IO pool based on IO size, deprecate io_unit_size
Current logic in transport buffers management uses io_unit_size to
calculate the number of buffers needed to process an IO request, size of
a single buffer is described by io_unit_size and iobuf selects either
small or large pool but never both for a single transport.
This change allows to select buffer size depending on IO size.
That helps to better memory utilization and to tune SPDK for better
BW utilization. This change is especially important for RDMA transport
where 1 large buffer instead of 16 small buffers shows much better BW
numbers and smaller latency.
io_unit_size parameter is deprecated. Many rdma tests are duplicated
for cases where small_iobuf_size == large_iobuf_size (the same as
IO with large payload which uses only large buffers) and for usual case
where a single large buffer is allocated instead of many small buffers
This commit contains a bug fix for iov iteration with DIF enabled -
iovpos was incremented when either we reached end of iov or consumed
all request's length. With this change iov len can be bigger than req's
length and we can increment iovpos incorrectly. This fix is part of the
commit due to other changes in this patch which are required to hit the
issue in UT.
The spdk_nvmf_transport structure was extended with 2 new fields
`large_bufsize` and `small_bufsize` to cache iobuf config and avoid
fetching it in IO path.
Default small/large cache sizes in TCP and RDMA transports are set to
UINT32_MAX to enable caching of both iobuf pools.

Change-Id: Ic7124a13d7bfcef14dd842255fc5dd7391041c69
Signed-off-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/27692
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>