lib/nvmf: add ns preempt hostids list
Add initial handling for preempt-and-abort command which builds a list of
preempted hostids. This list was already being built for reservation log
notifications but we need to save it in struct spdk_nvmf_ns for later
processing by poll groups.

This introduces a new structure spdk_nvmf_reserv_preempt_abort_info which
will contain the hostids list and in future commits more state related to
preempt-and-abort. The structure is linked in to struct spdk_nvmf_ns as
a pointer to avoid growing the structure for every use case of
namespaces.  The hostid list requires 256 bytes.

Change-Id: I8c3daad1333ef3d18eac087bcd97846332f62bf6
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26672
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>