test/unit/nvmf: add preempt with no holder test
Add unit test for preempt command with no reservation holder.

This tests registers all hosts with the same rkey and then preempt will
unregister all of them.

For preempt-and-abort we make sure the hostid that issues the preempt
doesn't abort itself (i.e. is not in the preempted hostids list).

Change-Id: I3d7425154a900c75793d872b083ddaa151d167ea
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26674
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>