spdk_nvme_perf: support percent-based number-ios (-d <percent>%)
This change allows the --number-ios/-d parameter to accept an integer
percent value (e.g., -d 50%), which is interpreted as a percentage of
the namespace size. The code calculates the number of I/Os after
namespace registration, using the percent value if specified, otherwise
falling back to the absolute value.

Although this complicates the code slightly, it is worthwhile because it
provides needed functionality that simplifies running scripts.
Previously, scripts had to externally calculate the percent of the
namespace size and pass the result as an absolute value. Now, users can
specify a percent directly, making automation and test scripting easier
and less error-prone.

Change-Id: I91e9565bb8d7bb6357ca88db6d42a990f297a723
Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26424
Reviewed-by: Ankit Kumar <ankit.kumar@samsung.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>