lib/nvmf: add preempt-abort IO wait tracking
Add infrastructure for preempt-and-abort to build an IO waiting count in
struct spdk_nvmf_subsystem_pg_ns_info based on the number of outstanding IOs
on the preempted controllers.

This builds upon the previous work of serializing reservation requests
and buliding preempted hostids list.

This doesn't yet hold the preempt-and-abort response until the IO
waiting is complete, that will come in another commit.

Add a unit tests that exercises preempt-and-abort and validates basic
building of the io_waiting count.

Change-Id: I5b72eb8ca908eda378b46d6605e1149a233e1169
Signed-off-by: Joel Cunningham <joel.cunningham@oracle.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26678
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Community-CI: Mellanox Build Bot