lib/bdev: add bdev_resize notification
Sparsely used lib notify can be used to catch events of importance
for the caller.

One example usage would be to catch resize of bdevs from a remote like
NVMe-oF.

For right now this will be used in bdev_nvme tests for namespace
add/remove. In some cases fixed later in the series, frequent handling
of ns changes could cause a nsdata to be zeroed out for a namespace
that was inactive. Yet this triggered a bdev resize event to 0 size.

Change-Id: Iadbe4eaa0fdd9c18c84d9fb68de0ba7960d56ec8
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26474
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>