nvmf: handle reservation notification log page count rollover
According to the NVMe specification, the reservation notification
log page count must roll over to 1h when it reaches its maximum value
(FFFFFFFF_FFFFFFFFh).

Log Page Count (LPC): This is a 64-bit incrementing Reservation
Notification log page count, indicating a unique identifier
(modulo 64 bit) for this notification. The count starts at 0h following
a Controller Level Reset and is incremented for every event that causes
a reservation notification regardless of whether that notification is
added to the queue. If the value of this field is FFFFFFFF_FFFFFFFFh,
then the field is set to 1h when incremented (i.e., rolls over to 1h)
and a new log page is created.

Change-Id: Ia03be3a344eb695bc1a47a1a07a85f7b19a7fe46
Signed-off-by: Dor Deri <dor.deri@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26463
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@nvidia.com>