nvmf: set correct cntlid for reservation report
Previously, the Controller ID (cntlid) in the Reservation Report was
always set to 0xFFFF, which is only valid when the registrant of the
namespace is not associated with any controller.

Updated the logic for setting ctrlr_data.cntlid to align with the
NVMe specification.

As defined in the NVMe spec:
Controller ID (CNTLID): If a registrant of the namespace is associated
with a controller in the NVM subsystem, then this field contains the
controller ID (i.e., the value of the CNTLID field in the Identify
Controller data structure) of the controller that is associated with
the registrant whose host identifier is indicated in the Host identifier
field of this Registrant data structure. If a registrant of the namespace
is not associated with any controller in the NVM subsystem, then the
controller processing the command shall set this field to FFFFh.

This ensures accurate reporting of registered controllers in compliance
with NVMe specification.

Change-Id: I1d6b374433411c50103a8c1c674a0132e4ca9382
Signed-off-by: Dor Deri <dor.deri@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26473
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>