nvme: simplify wait for completion interface
All usages pass the controller’s mutex and use the same timeout value
from the controller’s options. The interface can therefore be simplified
to improve readability and reduce the chances of misuse.

Put adminq into the function name to make it clear it is not for general use.

Change-Id: I5cdbae1162eb7f0c73d7cc6cdf90522829ba9ae0
Signed-off-by: Jacek Kalwas <jacek.kalwas@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26498
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>