lib/nvmf: implement WCTEMP and CCTEMP
NVMe 1.3d, section 5.15.3, Figure 112, specifies for Identify
Controller fields WCTEMP and CCTEMP that:
"Implementations compliant to revision 1.2 or later of this
specification shall report a non-zero value in this field."

To conform to the NVMe specification, set WCTEMP to the
recommended value of 0157h (343 K, 70 deg C).
Set CCTEMP to a dummy value of 0175h (373 K, 100 deg C).

Change-Id: Id81841c1bd98c9ce6886781e6994d9395d2d57ce
Signed-off-by: Marcin Galecki <marcin.galecki@dell.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26523
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>