test/common: account for actually executed tests
Compared skipped_tests.txt to completions_table from spdk-ci and removed tests
that are actually executed at this time.

Note that this list contains tests that are skipped for various reasons. Allowing
for a bit of leeway in terms of synchronizing changes in spdk and spdk-ci.
Thus we do not check test execution with this list for actually executed tests,
as that would be counter productive to the itended goal of catching tests that
would be disabled by mistake.

Change-Id: Ie5aa7ea59280a843c139b15d53b336aac518fb28
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26545
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>